Created a ScoreKeeper System
When the player destroys an enemy he will be rewarded a number of score points. The number of scorepoints is serialized so that it can be easily modified when for example introducitng new types of enemies. On the other hand when the player dies the score will be reset back to zero.